+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/main/scala/com/azure/cosmos/spark/udf/GetFeedRangesForContainer.scala
@@ -18,7 +18,7 @@ class GetFeedRangesForContainer extends UDF2[Map[String, String], Option[Int], A
     val effectiveUserConfig = CosmosConfig.getEffectiveConfig(None, None, userProvidedConfig)
     var feedRanges = List[String]()
     val cosmosContainerConfig: CosmosContainerConfig =
-      CosmosContainerConfig.parseCosmosContainerConfig(effectiveUserConfig, None, None)
+      CosmosContainerConfig.parseCosmosContainerConfig(effectiveUserConfig)
     val readConfig = CosmosReadConfig.parseCosmosReadConfig(effectiveUserConfig)
     val cosmosClientConfig = CosmosClientConfiguration(
       effectiveUserConfig,