+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/test/scala/com/azure/cosmos/spark/BulkWriterITest.scala
@@ -31,11 +31,13 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemOverwrite, 5, bulkEnabled = true, bulkMaxPendingOperations = Some(900))
 
     val metricsPublisher = new TestOutputMetricsPublisher
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -71,10 +73,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemOverwrite, 5, bulkEnabled = true, bulkMaxPendingOperations = Some(900))
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -118,10 +122,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       cosmosClient.getDatabase(cosmosDatabase).getContainer(containerCreationResponse.getProperties.getId)
 
     try {
+      val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
       val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemOverwrite, 5, bulkEnabled = true, bulkMaxPendingOperations = Some(900))
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -183,11 +189,13 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       cosmosClient.getDatabase(cosmosDatabase).getContainer(containerCreationResponse.getProperties.getId)
 
     try {
+      val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
       val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemAppend, 5, bulkEnabled = true, bulkMaxPendingOperations = Some(900))
 
       val metricsPublisher = new TestOutputMetricsPublisher
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -238,11 +246,13 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemOverwrite, 5, bulkEnabled = true, bulkMaxPendingOperations = Some(900))
 
     val metricsPublisher = new TestOutputMetricsPublisher
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -279,6 +289,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkDeleter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       deleteConfig,
       DiagnosticsConfig(),
@@ -304,10 +315,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemOverwrite, 5, bulkEnabled = true, bulkMaxPendingOperations = Some(900))
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -334,6 +347,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkUpdater = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -372,6 +386,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkDeleter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       deleteConfig,
       DiagnosticsConfig(),
@@ -392,9 +407,11 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val container = getContainer
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(ItemWriteStrategy.ItemAppend, maxRetryCount = 5, bulkEnabled = true, Some(900))
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -426,6 +443,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwriteIfNotModified,
       5,
@@ -436,6 +454,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val metricsPublisher = new TestOutputMetricsPublisher
     var bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -462,6 +481,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -499,6 +519,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val thirdWriteId = UUID.randomUUID().toString
     bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -546,6 +567,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
     val strippedPartitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
 
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -556,6 +578,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val metricsPublisher = new TestOutputMetricsPublisher
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -598,6 +621,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(
           columnConfigsMap,
           container,
+          containerConfig,
           partitionKeyDefinition,
           None,
           metricsPublisher)
@@ -673,6 +697,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
     val partitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -682,6 +707,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -709,7 +735,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         columnConfigsMap += "newItemInPropArray" -> CosmosPatchColumnConfig(
           "newItemInPropArray", operationType, "/propArray/0", isRawJson = false)
 
-        val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+        val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
         val patchPartialUpdateItem = CosmosPatchTestHelper.getPatchItemWithSchema(id, partitionKeyPath, partialUpdateSchema)
 
         operationType match {
@@ -751,6 +777,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
     val strippedPartitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -760,6 +787,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -794,7 +822,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
           field.getKey, CosmosPatchOperationTypes.Increment, s"/${field.getKey}", isRawJson = false)
       })
 
-      val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+      val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
 
       bulkWriterForPatch.scheduleWrite(partitionKey, patchPartialUpdateItem)
       bulkWriterForPatch.flushAndClose()
@@ -823,7 +851,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         field.getKey, CosmosPatchOperationTypes.Increment, s"/${field.getKey}", isRawJson = false)
     })
 
-    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
 
     bulkWriterForPatch.scheduleWrite(partitionKey, patchPartialUpdateItem)
     bulkWriterForPatch.flushAndClose()
@@ -833,6 +861,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val container = getContainer
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -842,6 +871,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -875,7 +905,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     columnConfigsMap += parent2PropertyName -> CosmosPatchColumnConfig(
       parent2PropertyName, CosmosPatchOperationTypes.Add, s"/family/parent2", isRawJson = false)
 
-    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
     bulkWriterForPatch.scheduleWrite(partitionKey, partialUpdateNode)
     bulkWriterForPatch.flushAndClose()
 
@@ -892,6 +922,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val container = getContainer
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -901,6 +932,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -944,7 +976,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     columnConfigsMap += "numberNode" -> CosmosPatchColumnConfig(
       "numberNode", CosmosPatchOperationTypes.Add, s"/family/numberNode", isRawJson = true)
 
-    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
     bulkWriterForPatch.scheduleWrite(partitionKey, partialUpdateNode)
     bulkWriterForPatch.flushAndClose()
 
@@ -968,6 +1000,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val container = getContainer
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -977,6 +1010,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -1010,7 +1044,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     columnConfigsMap += parent1PropertyName -> CosmosPatchColumnConfig(
       parent1PropertyName, CosmosPatchOperationTypes.Replace, s"/family/parent1", isRawJson = false)
 
-    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
     bulkWriterForPatch.scheduleWrite(partitionKey, partialUpdateNode)
     bulkWriterForPatch.flushAndClose()
 
@@ -1028,6 +1062,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
     val strippedPartitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -1037,6 +1072,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -1076,7 +1112,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         field.getKey, CosmosPatchOperationTypes.Set, s"/${field.getKey}", isRawJson = false)
     })
 
-    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
 
     bulkWriterForPatch.scheduleWrite(partitionKey, patchPartialUpdateItem)
     bulkWriterForPatch.flushAndClose()
@@ -1103,6 +1139,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         cosmosClient.getDatabase(cosmosDatabase).getContainer(containerCreationResponse.getProperties.getId)
 
       try {
+        val containerConfig = CosmosContainerConfig(createdContainer.getDatabase.getId, createdContainer.getId, None)
         val writeConfig = CosmosWriteConfig(
           ItemWriteStrategy.ItemOverwrite,
           5,
@@ -1112,6 +1149,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
         val bulkWriter = new BulkWriter(
           createdContainer,
+          containerConfig,
           partitionKeyDefinition,
           writeConfig,
           DiagnosticsConfig(),
@@ -1145,7 +1183,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
             field.getKey, CosmosPatchOperationTypes.Set, s"/${field.getKey}", isRawJson = false)
         })
 
-        val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, createdContainer, partitionKeyDefinition)
+        val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, createdContainer, containerConfig, partitionKeyDefinition)
 
         bulkWriterForPatch.scheduleWrite(partitionKey, patchPartialUpdateItem)
         bulkWriterForPatch.flushAndClose()
@@ -1163,6 +1201,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
     val strippedPartitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val writeConfig = CosmosWriteConfig(
       ItemWriteStrategy.ItemOverwrite,
       5,
@@ -1172,6 +1211,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
     val bulkWriter = new BulkWriter(
       container,
+      containerConfig,
       partitionKeyDefinition,
       writeConfig,
       DiagnosticsConfig(),
@@ -1210,6 +1250,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       CosmosPatchTestHelper.getBulkWriterForPatch(
         columnConfigsMap,
         container,
+        containerConfig,
         partitionKeyDefinition,
         Some(s"from c where c.propInt > ${Integer.MAX_VALUE}")) // using a always false condition
 
@@ -1231,6 +1272,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
   "Bulk Writer" should "throw exception if no valid operations are included in patch operation" in {
     val container = getContainer
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
     val containerProperties = container.read().block().getProperties
     val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
     val partitionKeyPath = StringUtils.join(partitionKeyDefinition.getPaths, "")
@@ -1247,7 +1289,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         field.name, CosmosPatchOperationTypes.Set, s"/${field.name}", isRawJson = false)
     })
 
-    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, partitionKeyDefinition)
+    val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatch(columnConfigsMap, container, containerConfig, partitionKeyDefinition)
     val patchPartialUpdateItem = CosmosPatchTestHelper.getPatchItemWithSchema(id, partitionKeyPath, partialUpdateSchema)
 
     try {
@@ -1264,8 +1306,9 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val container = getContainer
       val containerProperties = container.read().block().getProperties
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+      val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
 
-      val metricsPublisher = new TestOutputMetricsPublisher
+    val metricsPublisher = new TestOutputMetricsPublisher
       // if the item does not exists, patchBulkUpdate essentially will create those items
       // Validate that patchBulkUpdate can create items successfully
       val writeConfig = CosmosWriteConfig(
@@ -1277,6 +1320,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -1313,7 +1357,8 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val containerProperties = container.read().block().getProperties
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
 
-      val writeConfig = CosmosWriteConfig(
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
+    val writeConfig = CosmosWriteConfig(
           ItemWriteStrategy.ItemBulkUpdate,
           5,
           bulkEnabled = true,
@@ -1322,6 +1367,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -1359,7 +1405,8 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
       val strippedPartitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
 
-      val writeConfig = CosmosWriteConfig(
+    val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
+    val writeConfig = CosmosWriteConfig(
           ItemWriteStrategy.ItemOverwrite,
           5,
           bulkEnabled = true,
@@ -1368,6 +1415,7 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -1397,7 +1445,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
           columnConfigsMap += field.getKey -> CosmosPatchColumnConfig(field.getKey, CosmosPatchOperationTypes.Set, s"/${field.getKey}", isRawJson = false)
       })
 
-      val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(columnConfigsMap, container, partitionKeyDefinition)
+      val bulkWriterForPatch =
+        CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(
+          columnConfigsMap,
+          container,
+          containerConfig,
+          partitionKeyDefinition)
 
       bulkWriterForPatch.scheduleWrite(partitionKey, patchPartialUpdateItem)
       bulkWriterForPatch.flushAndClose()
@@ -1433,15 +1486,17 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val containerProperties = container.read().block().getProperties
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
       val partitionKeyPath = CosmosPatchTestHelper.getStrippedPartitionKeyPath(partitionKeyDefinition)
+      val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
       val writeConfig = CosmosWriteConfig(
-          ItemWriteStrategy.ItemOverwrite,
-          5,
-          bulkEnabled = true,
-          bulkMaxPendingOperations = Some(900)
-      )
+            ItemWriteStrategy.ItemOverwrite,
+            5,
+            bulkEnabled = true,
+            bulkMaxPendingOperations = Some(900)
+        )
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -1465,7 +1520,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       columnConfigsMap += "newItemInPropArray" -> CosmosPatchColumnConfig(
           "newItemInPropArray", CosmosPatchOperationTypes.Set, "/propArray/0", isRawJson = false)
 
-      val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(columnConfigsMap, container, partitionKeyDefinition)
+      val bulkWriterForPatch =
+        CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(
+          columnConfigsMap,
+          container,
+          containerConfig,
+          partitionKeyDefinition)
       val patchPartialUpdateItem = CosmosPatchTestHelper.getPatchItemWithSchema(id, partitionKeyPath, partialUpdateSchema)
 
       bulkWriterForPatch.scheduleWrite(partitionKey, patchPartialUpdateItem)
@@ -1483,15 +1543,17 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val container = getContainer
       val containerProperties = container.read().block().getProperties
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+      val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
       val writeConfig = CosmosWriteConfig(
-          ItemWriteStrategy.ItemOverwrite,
-          5,
-          bulkEnabled = true,
-          bulkMaxPendingOperations = Some(900)
-      )
+            ItemWriteStrategy.ItemOverwrite,
+            5,
+            bulkEnabled = true,
+            bulkMaxPendingOperations = Some(900)
+        )
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -1524,7 +1586,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       columnConfigsMap += parent2PropertyName -> CosmosPatchColumnConfig(
           parent2PropertyName, CosmosPatchOperationTypes.Set, s"/family/parent2", isRawJson = false)
 
-      val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(columnConfigsMap, container, partitionKeyDefinition)
+      val bulkWriterForPatch =
+        CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(
+          columnConfigsMap,
+          container,
+          containerConfig,
+          partitionKeyDefinition)
       bulkWriterForPatch.scheduleWrite(partitionKey, partialUpdateNode)
       bulkWriterForPatch.flushAndClose()
 
@@ -1541,15 +1608,17 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
       val container = getContainer
       val containerProperties = container.read().block().getProperties
       val partitionKeyDefinition = containerProperties.getPartitionKeyDefinition
+      val containerConfig = CosmosContainerConfig(container.getDatabase.getId, container.getId, None)
       val writeConfig = CosmosWriteConfig(
-          ItemWriteStrategy.ItemOverwrite,
-          5,
-          bulkEnabled = true,
-          bulkMaxPendingOperations = Some(900)
-      )
+            ItemWriteStrategy.ItemOverwrite,
+            5,
+            bulkEnabled = true,
+            bulkMaxPendingOperations = Some(900)
+        )
 
       val bulkWriter = new BulkWriter(
         container,
+        containerConfig,
         partitionKeyDefinition,
         writeConfig,
         DiagnosticsConfig(),
@@ -1593,7 +1662,12 @@ class BulkWriterITest extends IntegrationSpec with CosmosClient with AutoCleanab
         columnConfigsMap += "numberNode" -> CosmosPatchColumnConfig(
             "numberNode", CosmosPatchOperationTypes.Set, s"/family/numberNode", isRawJson = true)
 
-        val bulkWriterForPatch = CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(columnConfigsMap, container, partitionKeyDefinition)
+        val bulkWriterForPatch =
+          CosmosPatchTestHelper.getBulkWriterForPatchBulkUpdate(
+            columnConfigsMap,
+            container,
+            containerConfig,
+            partitionKeyDefinition)
         bulkWriterForPatch.scheduleWrite(partitionKey, partialUpdateNode)
         bulkWriterForPatch.flushAndClose()